Fix : 파일 위치 변경, 엔티티 순서 변경
